requirements.txt
argparse==1.3.0
dj-database-url==0.3.0
django-extensions==1.5.2
django-filter==0.9.2
Django==1.8.2
djangorestframework==3.1.1
gunicorn==19.3
Markdown==2.6.1
newrelic==2.50.0.39
Pillow==2.8.1
psycopg2==2.6
pytz==2015.2
six==1.9.0
whitenoise==1.0.6
wsgiref==0.1.2
beautifulsoup4==4.3.2
lxml==3.4.4
django-haystack==2.4.0
elasticsearch==1.6.0
-e git+https://github.com/edsu/microdata.git#egg=microdata